fix: bring the Python examples up to Python 3
- 13 tutorials plus cmake/passthru.py used Python 2 print statements, so
  they raised SyntaxError before running a line. Trailing-comma prints
  become print(..., end=' ').
- cmake/passthru.py also used `except EnvironmentError, e`; EnvironmentError
  is an alias of OSError.
- long() -> int() in the four mcmain ising.py tutorials. Python 3 has a
  single unbounded int, so these raised NameError immediately.
- .iteritems() -> .items() in three tutorials.
- code-09 and ngs/7 called `sim.paramters`; the attribute is `parameters`.
  This was masked until obj.save(archive) dispatch started working again.

tutorials/code-01-python/ising-skeleton.py is deliberately left alone: its
loop bodies are `#...` placeholders for the reader to fill in, with the
worked version in solution/, so it is not meant to parse.

+++ tutorials/code-08-mcmain-python/ising.py
@@ -30,8 +30,8 @@ class sim:
 
         self.length = int(self.parameters['L'])
         self.sweeps = 0
-        self.thermalization_sweeps = long(self.parameters['THERMALIZATION'])
-        self.total_sweeps = long(self.parameters['SWEEPS'])
+        self.thermalization_sweeps = int(self.parameters['THERMALIZATION'])
+        self.total_sweeps = int(self.parameters['SWEEPS'])
         self.beta = 1. / float(self.parameters['T'])
         self.spins = np.array([(-x if self.random() < 0.5 else x) for x in np.ones(self.length)])
