+++ tutorials/code-09-mcmain-python-hybrid/ising.py
@@ -25,8 +25,8 @@ class sim(ngs.mcbase):
 
         self.length = int(self.parameters['L'])
         self.sweeps = 0
-        self.thermalization_sweeps = long(self.parameters['THERMALIZATION'])
-        self.total_sweeps = long(self.parameters['SWEEPS'])
+        self.thermalization_sweeps = int(self.parameters['THERMALIZATION'])
+        self.total_sweeps = int(self.parameters['SWEEPS'])
         self.beta = 1. / float(self.parameters['T'])
         self.spins = np.array([(-x if self.random() < 0.5 else x) for x in np.ones(self.length)])
